requirements.txt
pillow
typer
